+++ Modular_Arithmetic/MatrixModulo.c
@@ -8,16 +8,16 @@
 /**
  * Prints a matrix modulo n.
  *
- * @param a the matrix to print - M: k x m.
+ * @param a the matrix to prlong long int - M: k x m.
  * @param n the module value.
  */
-void printMatrixModulo(matrix *a, int n) {
+void printMatrixModulo(matrix *a, long long int n) {
     assert(a->n > 0);
     assert(a->m > 0);
 
-    for (int i = 0; i < a->n; ++i) {
-        for (int j = 0; j < a->m; ++j) {
-            printf("%d ", mod((int) a->matrix[i][j], n));
+    for (long long int i = 0; i < a->n; ++i) {
+        for (long long int j = 0; j < a->m; ++j) {
+            printf("%lld ", mod((int) a->matrix[i][j], n));
         }
         printf("\n");
     }
@@ -30,12 +30,12 @@ void printMatrixModulo(matrix *a, int n) {
  * @param a the matrix - M: k x m.
  * @return 1 if the matrix has all integer elements, 0 otherwise.
  */
-int isIntegerMatrix(matrix *a) {
+long long int isIntegerMatrix(matrix *a) {
     assert(a->n > 0);
     assert(a->m > 0);
 
-    for (int i = 0; i < a->n; ++i) {
-        for (int j = 0; j < a->m; ++j) {
+    for (long long int i = 0; i < a->n; ++i) {
+        for (long long int j = 0; j < a->m; ++j) {
             if (a->matrix[i][j] != (int) a->matrix[i][j]) {
                 return 0;
             }
@@ -53,7 +53,7 @@ int isIntegerMatrix(matrix *a) {
  * @param modMatrix the matrix modulo n - M: k x m.
  * @param n the module value.
  */
-void modularMatrix(matrix *a, matrix *modMatrix, int n) {
+void modularMatrix(matrix *a, matrix *modMatrix, long long int n) {
     assert(a->n > 0);
     assert(a->m > 0);
     assert(isIntegerMatrix(a) && "the matrix must have all element integer");
@@ -65,8 +65,8 @@ void modularMatrix(matrix *a, matrix *modMatrix, int n) {
     assert(modMatrix->n == a->n);
     assert(modMatrix->m == a->m);
 
-    for (int i = 0; i < a->n; ++i) {
-        for (int j = 0; j < a->m; ++j) {
+    for (long long int i = 0; i < a->n; ++i) {
+        for (long long int j = 0; j < a->m; ++j) {
             modMatrix->matrix[i][j] = mod((int) a->matrix[i][j], n);
             modMatrix->matrix[i][j] = modMatrix->matrix[i][j] < 0 ? modularReduction((int) modMatrix->matrix[i][j], n) : modMatrix->matrix[i][j];
         }
@@ -81,7 +81,7 @@ void modularMatrix(matrix *a, matrix *modMatrix, int n) {
  * @param inv the inverse of the matrix: [a]^-1 - M: n x n.
  * @param n the module value.
  */
-void inverseMatrixModulo(matrix *a, matrix *inv, int n) {
+void inverseMatrixModulo(matrix *a, matrix *inv, long long int n) {
     assert(a->n > 0);
     assert(a->m > 0);
     assert(a->m == a->n);
@@ -95,12 +95,12 @@ void inverseMatrixModulo(matrix *a, matrix *inv, int n) {
     assert(inv->m == a->m);
 
     //Inverse of the determinant.
-    int invDet = modularInverse((int) determinantMatrix(a), n);
+    long long int invDet = modularInverse((int) determinantMatrix(a), n);
     //Temporary matrix.
     matrix *tmp = createMatrix(a->n, a->m);
 
-    for (int i = 0; i < a->n; ++i) {
-        for (int j = 0; j < a->n; ++j) {
+    for (long long int i = 0; i < a->n; ++i) {
+        for (long long int j = 0; j < a->n; ++j) {
             tmp->matrix[j][i] = product(invDet, (int) cofactor(a, i, j), n);
             tmp->matrix[i][j] = tmp->matrix[i][j] < 0 ? modularReduction((int) tmp->matrix[i][j], n) : tmp->matrix[i][j];
         }
@@ -120,7 +120,7 @@ void inverseMatrixModulo(matrix *a, matrix *inv, int n) {
  * @param res the result: [a] + [b] - M: k x m.
  * @param n the module value.
  */
-void sumMatrixModulo(matrix *a, matrix *b, matrix *res, int n) {
+void sumMatrixModulo(matrix *a, matrix *b, matrix *res, long long int n) {
     assert(a->n > 0);
     assert(a->m > 0);
     assert(b->n > 0);
@@ -136,8 +136,8 @@ void sumMatrixModulo(matrix *a, matrix *b, matrix *res, int n) {
     assert(res->n == a->n);
     assert(res->m == a->m);
 
-    for (int i = 0; i < a->n; ++i) {
-        for (int j = 0; j < a->m; ++j) {
+    for (long long int i = 0; i < a->n; ++i) {
+        for (long long int j = 0; j < a->m; ++j) {
             res->matrix[i][j] = sum((int) a->matrix[i][j], (int) b->matrix[i][j], n);
             res->matrix[i][j] = res->matrix[i][j] < 0 ? modularReduction((int) res->matrix[i][j], n) : res->matrix[i][j];
         }
@@ -153,7 +153,7 @@ void sumMatrixModulo(matrix *a, matrix *b, matrix *res, int n) {
  * @param res the result: [a] - [b] - M: k x m.
  * @param n the module value.
  */
-void subMatrixModulo(matrix *a, matrix *b, matrix *res, int n) {
+void subMatrixModulo(matrix *a, matrix *b, matrix *res, long long int n) {
     assert(a->n > 0);
     assert(a->m > 0);
     assert(b->n > 0);
@@ -169,8 +169,8 @@ void subMatrixModulo(matrix *a, matrix *b, matrix *res, int n) {
     assert(res->n == a->n);
     assert(res->m == a->m);
 
-    for (int i = 0; i < a->n; ++i) {
-        for (int j = 0; j < a->m; ++j) {
+    for (long long int i = 0; i < a->n; ++i) {
+        for (long long int j = 0; j < a->m; ++j) {
             res->matrix[i][j] = sum((int) a->matrix[i][j], modularReduction((int) b->matrix[i][j], n), n);
             res->matrix[i][j] = res->matrix[i][j] < 0 ? modularReduction((int) res->matrix[i][j], n) : res->matrix[i][j];
         }
@@ -186,7 +186,7 @@ void subMatrixModulo(matrix *a, matrix *b, matrix *res, int n) {
  * @param res the result: scalar * [a] - M: k x m.
  * @param n the module value.
  */
-void scalarProductModulo(int scalar, matrix *a, matrix *res, int n) {
+void scalarProductModulo(long long int scalar, matrix *a, matrix *res, long long int n) {
     assert(a->n > 0);
     assert(a->m > 0);
     assert(isIntegerMatrix(a) && "the matrix must have all element integer");
@@ -197,8 +197,8 @@ void scalarProductModulo(int scalar, matrix *a, matrix *res, int n) {
     assert(res->n == a->n);
     assert(res->m == a->m);
 
-    for (int i = 0; i < a->n; ++i) {
-        for (int j = 0; j < a->m; ++j) {
+    for (long long int i = 0; i < a->n; ++i) {
+        for (long long int j = 0; j < a->m; ++j) {
             res->matrix[i][j] = product(scalar, (int) a->matrix[i][j], n);
             res->matrix[i][j] = res->matrix[i][j] < 0 ? modularReduction((int) res->matrix[i][j], n) : res->matrix[i][j];
         }
@@ -214,7 +214,7 @@ void scalarProductModulo(int scalar, matrix *a, matrix *res, int n) {
  * @param res the result: [a] x [b] - M: k x m.
  * @param n the module value.
  */
-void productMatrixModulo(matrix *a, matrix *b, matrix *res, int n) {
+void productMatrixModulo(matrix *a, matrix *b, matrix *res, long long int n) {
     assert(a->n > 0);
     assert(a->m > 0);
     assert(b->n > 0);
@@ -230,12 +230,12 @@ void productMatrixModulo(matrix *a, matrix *b, matrix *res, int n) {
     assert(res->m == b->m);
 
     //Temporary variable used to save the product in position (i; j).
-    int tmp = 0;
+    long long int tmp = 0;
 
-    for (int i = 0; i < a->n; ++i) {
-        for (int j = 0; j < b->m; ++j) {
+    for (long long int i = 0; i < a->n; ++i) {
+        for (long long int j = 0; j < b->m; ++j) {
             res->matrix[i][j] = 0;
-            for (int k = 0; k < a->m; ++k) {
+            for (long long int k = 0; k < a->m; ++k) {
                 tmp = sum(tmp, product((int) a->matrix[i][j], (int) b->matrix[i][j], n), n);
             }
             res->matrix[i][j] = tmp;
@@ -253,7 +253,7 @@ void productMatrixModulo(matrix *a, matrix *b, matrix *res, int n) {
  * @param res the result: [a]^k - M: m x m.
  * @param n the module value.
  */
-void powerMatrixModulo(matrix *a, int k, matrix *res, int n) {
+void powerMatrixModulo(matrix *a, long long int k, matrix *res, long long int n) {
     assert(a->n > 0);
     assert(a->m > 0);
     assert(a->m == a->n);
@@ -274,7 +274,7 @@ void powerMatrixModulo(matrix *a, int k, matrix *res, int n) {
     assert(res->m == a->m);
 
     //compute the power
-    for (int i = 0; i < k; ++i) {
+    for (long long int i = 0; i < k; ++i) {
         productMatrixModulo(res, aTmp, tmp, n);
         copyMatrix(tmp, res);
     }
@@ -293,7 +293,7 @@ void powerMatrixModulo(matrix *a, int k, matrix *res, int n) {
  * @param res the result: [a] (x) [b] - M: k*p x m*q.
  * @param n the module value,
  */
-void kroneckerProductMatrixModulo(matrix *a, matrix *b, matrix *res, int n) {
+void kroneckerProductMatrixModulo(matrix *a, matrix *b, matrix *res, long long int n) {
     assert(a->n > 0);
     assert(a->m > 0);
     assert(b->n > 0);
@@ -307,8 +307,8 @@ void kroneckerProductMatrixModulo(matrix *a, matrix *b, matrix *res, int n) {
     assert(res->n == a->n * b->n);
     assert(res->m == a->m * b->m);
 
-    for (int i = 0; i < a->n * b->n; ++i) {
-        for (int j = 0; j < a->m * b->m; ++j) {
+    for (long long int i = 0; i < a->n * b->n; ++i) {
+        for (long long int j = 0; j < a->m * b->m; ++j) {
             res->matrix[i][j] = product((int) a->matrix[i / b->n][j / b->m], (int) b->matrix[i % b->n][j % b->m], n);
             res->matrix[i][j] = res->matrix[i][j] < 0 ? modularReduction((int) res->matrix[i][j], n) : res->matrix[i][j];
         }
